requirements.txt
flask==2.0.1
geopy
mysql-connector-python
